+++ portfolio/src/app/pages/error-page/error-page.component.ts
@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
 import { TranslatePipe } from '@ngx-translate/core';
-import { ContainerComponent } from '../../shared/container/container.component';
+import { ContainerComponent, FadeInDirective } from '../../shared';
 import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-error-page',
-  imports: [TranslatePipe, ContainerComponent],
+  imports: [TranslatePipe, ContainerComponent, FadeInDirective],
   templateUrl: './error-page.component.html',
   styleUrl: './error-page.component.scss',
   standalone: true
